+++ mkFit/MkFinder.cc
@@ -362,7 +362,7 @@ void MkFinder::SelectHitIndices(const LayerOfHits &layer_of_hits,
         // and issue prefetches at the same time.
         // Then enter vectorized loop to actually collect the hits in proper order.
 
-        for (int hi = L.m_phi_bin_infos[qi][pb].first; hi < L.m_phi_bin_infos[qi][pb].second; ++hi)
+        for (uint16_t hi = L.m_phi_bin_infos[qi][pb].first; hi < L.m_phi_bin_infos[qi][pb].second; ++hi)
         {
           // MT: Access into m_hit_zs and m_hit_phis is 1% run-time each.
